title

+++ database/migrations/2018_04_16_233029_add_maximum_qiandao_to_users_table.php
@@ -14,7 +14,7 @@ class AddMaximumQiandaoToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('maximum_qiandao')->default(0);
+            $table->smallInteger('maximum_qiandao')->default(0);
         });
     }
